backend_real/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-filter==24.3
djoser==2.2.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
